requirements_backup.txt
0x-contract-addresses==3.0.0
0x-contract-artifacts==3.0.0
0x-json-schemas==2.1.0
0x-order-utils==4.0.1
# Editable install with no version control (agent0==0.1.0)
-e /Users/xo/xo-core-reforged/agent0/webui
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
arweave-python-client==1.0.19
async-timeout==5.0.1
attrs==25.3.0
bandit==1.8.3
base58==2.1.1
bcrypt==4.3.0
bitarray==2.9.3
black==25.1.0
cbor2==5.6.5
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
codespell==2.4.1
cryptography==45.0.3
cytoolz==0.12.3
dacite==1.9.2
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.7.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.3.0
fabric==3.2.2
fastapi==0.111.0
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.2.0
frozenlist==1.6.2
h11==0.16.0
hexbytes==0.3.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
ipfshttpclient==0.8.0a2
isort==6.0.1
Jinja2==3.1.6
jsonschema==3.2.0
lru-dict==1.3.0
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
multiaddr==0.0.9
multidict==6.4.4
mypy_extensions==0.4.4
netaddr==1.3.0
nodeenv==1.9.1
orjson==3.10.18
packaging==25.0
paramiko==3.5.1
parsimonious==0.8.1
pathspec==0.12.1
pbr==6.1.1
platformdirs==4.3.8
pre_commit==4.2.0
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pyee==9.1.1
pyflakes==3.3.2
Pygments==2.19.1
pymerkle==2.0.2
PyNaCl==1.5.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2022.7.1
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
rlp==2.0.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
stevedore==5.4.1
stringcase==1.2.0
thirdweb-contract-wrappers==2.0.4
thirdweb-eth-account==0.6.6
thirdweb-sdk==3.1.1
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.29.0
uvloop==0.21.0
varint==1.0.2
virtualenv==20.31.2
watchfiles==1.0.5
web3==5.27.0
websockets==15.0.1
wrapt==1.17.2
# Editable install with no version control (xo-agent==0.1.0)
-e /Users/xo/xo-core-reforged/xo-agent
yamllint==1.37.1
yarl==1.20.0
zipp==3.22.0
